% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_zcf.R
\name{paged_zcf}
\alias{paged_zcf}
\title{Function for windmill template}
\usage{
paged_zcf(
  logo = pkg_resource("logo/ssi-logo.png"),
  front_img = "0",
  img_to_dark = TRUE,
  logo_to_white = FALSE,
  type = "short",
  ...
)
}
\arguments{
\item{logo}{Logo}

\item{front_img}{Front image}

\item{img_to_dark}{Darken image}

\item{logo_to_white}{Tranform logo color to white - logo must be SVG}

\item{...}{Arguments passed to pagedown::html_paged}

\item{other_css}{Add an other CSS}
}
\value{
A pagedown template
}
\description{
Function for windmill template
}
