% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_sszcf_tw}
\alias{theme_sszcf_tw}
\title{Title}
\usage{
theme_sszcf_tw(
  base_family = "Titillium Web",
  base_size = 15,
  plot_title_family = if (.Platform$OS.type == "windows") "Titillium Web" else
    "Titillium Web Bold",
  plot_title_size = 18,
  plot_title_face = "bold",
  plot_title_margin = 10,
  subtitle_family = if (.Platform$OS.type == "windows") "Titillium Web" else
    "Titillium Web Light",
  subtitle_size = 13,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_family = base_family,
  strip_text_size = 12,
  strip_text_face = "plain",
  caption_family = if (.Platform$OS.type == "windows") "Titillium Web" else
    "Titillium Web Light",
  caption_size = 9,
  caption_face = "plain",
  caption_margin = 10,
  axis_text_size = base_size,
  axis_title_family = base_family,
  axis_title_size = 15,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = margin(30, 30, 30, 30),
  grid_col = "#FF9900",
  grid = TRUE,
  axis_col = "#FF9900",
  axis = FALSE,
  ticks = FALSE
)
}
\arguments{
\item{base_family}{}

\item{base_size}{}

\item{plot_title_family}{}

\item{plot_title_size}{}

\item{plot_title_face}{}

\item{plot_title_margin}{}

\item{subtitle_family}{}

\item{subtitle_size}{}

\item{subtitle_face}{}

\item{subtitle_margin}{}

\item{strip_text_family}{}

\item{strip_text_size}{}

\item{strip_text_face}{}

\item{caption_family}{}

\item{caption_size}{}

\item{caption_face}{}

\item{caption_margin}{}

\item{axis_text_size}{}

\item{axis_title_family}{}

\item{axis_title_size}{}

\item{axis_title_face}{}

\item{axis_title_just}{}

\item{plot_margin}{}

\item{grid_col}{}

\item{grid}{}

\item{axis_col}{}

\item{axis}{}

\item{ticks}{}
}
\value{

}
\description{
Title
}
